+++ pygenstub.py
@@ -41,7 +41,7 @@ from pkgutil import get_loader, walk_packages
 from docutils.core import publish_doctree
 
 
-__version__ = "1.4.0.dev1"  # sig: str
+__version__ = "1.4.0"  # sig: str
 
 
 PY3 = sys.version_info >= (3, 0)